+++ packages/apollo-react-relay-duct-tape/src/__tests__/__generated__/hooksTestFragment.graphql.ts
@@ -4,13 +4,13 @@
 
 import { FragmentRefs } from "@graphitation/apollo-react-relay-duct-tape";
 export type hooksTestFragment = {
-    readonly id: string;
-    readonly name: string;
-    readonly __typename: "User";
-    readonly " $refType": "hooksTestFragment";
+  readonly id: string;
+  readonly name: string;
+  readonly __typename: "User";
+  readonly " $fragmentType": "hooksTestFragment";
 };
 export type hooksTestFragment$data = hooksTestFragment;
 export type hooksTestFragment$key = {
-    readonly " $data"?: hooksTestFragment$data | undefined;
-    readonly " $fragmentRefs": FragmentRefs<"hooksTestFragment">;
+  readonly " $data"?: hooksTestFragment$data | undefined;
+  readonly " $fragmentSpreads": FragmentRefs<"hooksTestFragment">;
 };