+++ packages/apollo-react-relay-duct-tape/src/storeObservation/__tests__/compiledHooks.test.tsx
@@ -1,5 +1,9 @@
 import React from "react";
-import { ApolloClient, defaultDataIdFromObject } from "@apollo/client";
+import {
+  ApolloClient,
+  WatchQueryFetchPolicy,
+  defaultDataIdFromObject,
+} from "@apollo/client";
 import {
   act,
   create as createTestRenderer,
@@ -78,12 +82,17 @@ const _ForwardPagination_fragment = graphql`
     conversationsForwardCount: { type: "Int!", defaultValue: 1 }
     conversationsAfterCursor: { type: "String!", defaultValue: "" }
     addExtra: { type: "Boolean!", defaultValue: false }
+    sortBy: {
+      type: "SortByInput"
+      defaultValue: { sortField: NAME, sortDirection: ASC }
+    }
   ) {
     petName
     avatarUrl(size: $avatarSize)
     conversations(
       first: $conversationsForwardCount
       after: $conversationsAfterCursor
+      sortBy: $sortBy
     ) @connection(key: "compiledHooks_user_conversations") {
       edges {
         node {
@@ -217,10 +226,11 @@ describe.each([
 
   const RootComponent: React.FC<{
     variables: compiledHooks_Root_executionQueryVariables;
+    fetchPolicy?: WatchQueryFetchPolicy;
   }> = (props) => {
     const result = useCompiledLazyLoadQuery(
       compiledHooks_Root_executionQuery_documents,
-      { variables: props.variables },
+      { variables: props.variables, fetchPolicy: props.fetchPolicy },
     );
     useLazyLoadQueryResult = result;
     return result.data ? (
@@ -379,6 +389,7 @@ describe.each([
       });
 
       it("fetches new data when variables change", async () => {
+        const watchQuerySpy = jest.spyOn(client, "watchQuery");
         act(() => {
           testRenderer.update(
             <ApolloReactRelayDuctTapeProvider client={client}>
@@ -401,6 +412,60 @@ describe.each([
             }),
           ),
         );
+        expect(watchQuerySpy).toHaveBeenCalledWith({
+          query:
+            compiledHooks_Root_executionQuery_documents.executionQueryDocument,
+          variables: {
+            userId: 21,
+            messagesBackwardCount: 1,
+            messagesBeforeCursor: "",
+          },
+          fetchPolicy: undefined,
+        });
+        expect(
+          client.cache.extract()[
+            typePolicies === typePoliciesWithDefaultApolloClientStoreKeys
+              ? "User:21"
+              : "21"
+          ],
+        ).toMatchSnapshot();
+      });
+
+      it("fetches new data when fetchPolicy changes", async () => {
+        const watchQuerySpy = jest.spyOn(client, "watchQuery");
+        act(() => {
+          testRenderer.update(
+            <ApolloReactRelayDuctTapeProvider client={client}>
+              <ErrorBoundary>
+                <RootComponent
+                  variables={{
+                    userId: 21,
+                    messagesBackwardCount: 1,
+                    messagesBeforeCursor: "",
+                  }}
+                  fetchPolicy="network-only"
+                />
+              </ErrorBoundary>
+            </ApolloReactRelayDuctTapeProvider>,
+          );
+        });
+        await act(() =>
+          client.mock.resolveMostRecentOperation((operation) =>
+            MockPayloadGenerator.generate(operation, {
+              User: () => ({ id: operation.request.variables.userId }),
+            }),
+          ),
+        );
+        expect(watchQuerySpy).toHaveBeenCalledWith({
+          query:
+            compiledHooks_Root_executionQuery_documents.executionQueryDocument,
+          variables: {
+            userId: 21,
+            messagesBackwardCount: 1,
+            messagesBeforeCursor: "",
+          },
+          fetchPolicy: "network-only",
+        });
         expect(
           client.cache.extract()[
             typePolicies === typePoliciesWithDefaultApolloClientStoreKeys