+++ packages/apollo-react-relay-duct-tape/src/storeObservation/compiledHooks/useCompiledLazyLoadQuery.ts
@@ -5,7 +5,11 @@ import { useDeepCompareMemoize } from "./useDeepCompareMemoize";
 import { useForceUpdate } from "./useForceUpdate";
 import { useOverridenOrDefaultApolloClient } from "../../useOverridenOrDefaultApolloClient";
 
-import type { ObservableQuery, ApolloClient } from "@apollo/client";
+import type {
+  ObservableQuery,
+  ApolloClient,
+  WatchQueryFetchPolicy,
+} from "@apollo/client";
 import type { DocumentNode } from "graphql";
 import type { CompiledArtefactModule } from "@graphitation/apollo-react-relay-duct-tape-compiler";
 
@@ -53,6 +57,7 @@ function useExecutionQuery(
   client: ApolloClient<unknown>,
   executionQueryDocument: DocumentNode,
   variables: Record<string, unknown>,
+  fetchPolicy: WatchQueryFetchPolicy | undefined,
 ): [loading: boolean, error?: Error] {
   const forceUpdate = useForceUpdate();
   const execution = useRef(new ExecutionQueryHandler(() => forceUpdate()));
@@ -62,13 +67,14 @@ function useExecutionQuery(
         client.watchQuery({
           query: executionQueryDocument,
           variables,
+          fetchPolicy,
         }),
       );
     }
     return () => {
       execution.current.reset();
     };
-  }, [executionQueryDocument, variables]);
+  }, [executionQueryDocument, variables, fetchPolicy]);
   return execution.current.status;
 }
 
@@ -82,7 +88,10 @@ function useExecutionQuery(
  */
 export function useCompiledLazyLoadQuery(
   documents: CompiledArtefactModule,
-  options: { variables: Record<string, unknown> },
+  options: {
+    variables: Record<string, unknown>;
+    fetchPolicy?: WatchQueryFetchPolicy;
+  },
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
 ): { data?: Record<string, any>; error?: Error } {
   const { watchQueryDocument } = documents;
@@ -104,6 +113,7 @@ export function useCompiledLazyLoadQuery(
     client,
     executionQueryDocument,
     variables,
+    options.fetchPolicy,
   );
   // ...then fetch/watch data for only the calling component...
   const { data } = useApolloQuery(watchQueryDocument, {