+++ packages/embedded-document-artefact-loader/src/webpack.ts
@@ -3,12 +3,23 @@ import { SourceMapGenerator } from "source-map-js";
 import { transform } from "./transform";
 import { applySourceMap } from "./source-map-utils";
 
-const webpackLoader: LoaderDefinitionFunction = function (
+type Options = {
+  artifactDirectory?: string;
+};
+
+const webpackLoader: LoaderDefinitionFunction<Options> = function (
   source,
   inputSourceMap,
   _additionalData,
 ) {
   const callback = this.async();
+
+  // Using query instead of getOptions to support webpack 4
+  // and the loader-runner doesn't support getOptions (yet)
+  const options =
+    typeof this.query !== "string" && this.query.artifactDirectory
+      ? { artifactDirectory: this.query.artifactDirectory }
+      : undefined;
 
   let sourceMap: SourceMapGenerator | undefined;
   if (this.sourceMap) {
@@ -18,7 +29,7 @@ const webpackLoader: LoaderDefinitionFunction = function (
     sourceMap.setSourceContent(this.resourcePath, source);
   }
 
-  const transformed = transform(source, this.resourcePath, sourceMap);
+  const transformed = transform(source, this.resourcePath, sourceMap, options);
 
   if (transformed && sourceMap && inputSourceMap) {
     callback(